# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "News Aggregator API",
    "version": "1.0.0",
    "description": "Public API for aggregating news from multiple providers. This API returns curated news articles for users.",
    "contact": {
      "name": "News Team",
      "email": "shanmuka089@gmail.com",
      "url": "https://example.com/support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/news-aggregator/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080/news-aggregator/api/v1",
      "description": "Staging"
    }
  ],
...
  "paths": {
...
    "/news": {
      "get": {
        "tags": [
          "News Articles"
        ],
        "summary": "Get News Articles",
        "description": "Retrieves news articles based on user preferences",
        "requestBody": {},
        "responses": {
          "200": {
            "description": "News articles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_BAD_REQUEST",
                      "message": "Validation failed",
                      "errorMessage": "Invalid input data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_UNAUTHORIZED",
                      "message": "Unauthorized access",
                      "errorMessage": "You must be logged in to access this resource"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_FORBIDDEN",
                      "message": "Access denied",
                      "errorMessage": "You do not have permission to perform this action"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_INTERNAL_SERVER_ERROR",
                      "message": "An unexpected error occurred",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_SERVICE_UNAVAILABLE",
                      "message": "Service is currently unavailable",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [],
        "security": []
      }
    },
    "/news/{category}": {
      "get": {
        "tags": [
          "News Articles"
        ],
        "summary": "Get News Articles by Category",
        "description": "Retrieves news articles based on a specific category",
        "requestBody": {},
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
            "type": "string",
            "default": "Sports"
            },
            "description": "Category of news articles to retrieve"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en"
            },
            "description": "Language of the news articles (default is 'en')"
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "in"
            },
            "description": "Country of the news articles (default is 'in')"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all"
            },
            "description": "Source of the news articles (default is 'all')"
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Cricket"
            },
            "description": "Topic of the news articles (default is 'Cricket')"
          }
        ],
        "responses": {
          "200": {
            "description": "News articles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_BAD_REQUEST",
                      "message": "Validation failed",
                      "errorMessage": "Invalid input data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_UNAUTHORIZED",
                      "message": "Unauthorized access",
                      "errorMessage": "You must be logged in to access this resource"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_FORBIDDEN",
                      "message": "Access denied",
                      "errorMessage": "You do not have permission to perform this action"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_INTERNAL_SERVER_ERROR",
                      "message": "An unexpected error occurred",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_SERVICE_UNAVAILABLE",
                      "message": "Service is currently unavailable",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceDto"
          }
        }
      },
...
      "NewsResponseDto": {
        "type": "object",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleDto"
            }
          }
        }
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Error code representing the type of error",
            "example": "NEWS_API_CLIENT_BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "description": "A human-readable message providing more details about the error",
            "example": "Validation failed"
          },
          "errorMessage": {
            "type": "string",
            "description": "Detailed error message, if available",
            "example": "Invalid input data"
          }
        }
      },
...
      "SourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  }
}